requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Text Processing
sentencepiece>=0.1.99
transformers>=4.30.0
tokenizers>=0.13.3

# Vision and Diffusion
diffusers>=0.20.0
accelerate>=0.20.0
safetensors>=0.3.1
Pillow>=9.5.0
opencv-python>=4.7.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
encodec>=0.1.1

# Training Infrastructure
pytorch-lightning>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Data and Utilities
datasets>=2.12.0
pandas>=2.0.0
tqdm>=4.65.0
einops>=0.6.1
timm>=0.9.2

# Evaluation
scikit-learn>=1.2.2
torchmetrics>=0.11.4
pycocotools>=2.0.6

# API and Serving
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
redis>=4.5.0

# Development
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0
